unpapered_backend/requirements.txt
fastapi==0.110.0
PyPDF2==3.0.1
langchain==0.1.13
google==3.0.0
faiss-cpu==1.8.0
python-multipart==0.0.9
langchain-google-genai==0.0.11
python-dotenv
uvicorn[standard]
